requirements.txt
asgiref==3.6.0
Django==4.1.7
Pillow==9.4.0
sqlparse==0.4.3
django-allauth==0.54.0
sorl-thumbnail==12.9.0
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
pytils==0.4.1
python-dotenv==1.0.0
psycopg2-binary==2.9.5
django-modeltranslation==0.18.10